remove nvim folder; update install script for lazy.nvim

+++ lazy.nvim/install.sh
@@ -5,15 +5,15 @@ source ./functions.sh
 echo "[nvim] install python3-neovim and libfuse2 (for appimages)"
 sudo apt-get install python3-neovim libfuse2 -y &> /dev/null
 echo "[nvim] download nvim appimage"
-curl -sLO https://github.com/neovim/neovim/releases/latest/download/nvim.appimage &> /dev/null
-chmod u+x nvim.appimage
+curl -o /tmp/nvim -sL https://github.com/neovim/neovim/releases/latest/download/nvim.appimage &> /dev/null
+chmod u+x /tmp/nvim
 
 echo "[nvim] Link nvim.desktop file"
-createSymlink ~/repos/dotfiles/nvim/nvim.desktop /usr/share/applications/nvim.desktop
+createSymlink ~/repos/dotfiles/lazy.nvim/nvim.desktop /usr/share/applications/nvim.desktop
 
 binDirectory=/usr/local/bin/nvim
 echo "[nvim] nvim moved to \"$binDirectory\""
-sudo mv nvim.appimage $binDirectory
+sudo mv /tmp/nvim $binDirectory
 
 echo "[nvim] Link nvim folder on .config"
 createSymlink ~/repos/dotfiles/lazy.nvim ~/.config/nvim
@@ -25,8 +25,12 @@ mkdir -p ~/.config/nvim/{backup_files,swap_files,undo_files}/
 echo "[nvim] Install plugins dependencies"
 sudo apt-get install ripgrep fd-find -y &> /dev/null
 
-echo "[nvim] Link fdfind to be executed as fd"
-ln -s $(which fdfind) ~/.local/bin/fd
+if [[ -f ~/.local/bin/fd ]]; then
+  echo "[nvim] fdfind already linked"
+else
+  echo "[nvim] Link fdfind to be executed as fd"
+  ln -s $(which fdfind) ~/.local/bin/fd
+fi
 
 echo "[nvim] Install Lazy nvim and plugins"
 # The bang in the command will make it wait until it's finished
